Preserve Dual State w/ LocalStorage
Fix all sorts of weird bugs. Now, no matter how often you refresh, or set multiple states, behavior is as expected. This one was tricky. Basically, the initial solution (existing on Payment.js still) works really well if its only monitoring for one variable. If it disappears in local storage, reset state variable as well, triggering a re-render. This doesn't work for two variables, because they keep triggering re-renders and infinite loop. Now, use a dummy variable solution, and only set state in useEffect with that dummy variable as a dependency. This way, we do it once, states are set, and proper information can be returned