Fix cache bug
Caught a weird bug. If you sign out from profile page, your mnemonic phrase is still stored in payment page. If you happened to have an amount filled in amount box, and press transact, it will still transact! To fix, created a async function that runs every transaction. Makes sure to check if value is still in localstorage or not